requirements.txt
asgiref==3.5.2
beautifulsoup4==4.9.3
crispy-bootstrap5==0.7
Django==4.1.1
django-bootstrap5==23.1
django-email==0.1.10
django-environ==0.4.5
django-fontawesome-6==1.0.0.0
django-google-maps==0.13.0
django-jquery==3.1.0
django-multipleselectfield==0.2.0.post2
django-phonenumber-field[phonenumberslite]==7.0.*
Jinja2==3.1.2
MarkupSafe==2.1.2
Pillow==9.2.0
psycopg2==2.9.5
soupsieve==2.4
whitenoise==6.4.0